Extract non-DEM atlas entry factory
